+++ app/src/main/java/com/example/dmitry/a1c_client/android/IncomeTaskActivity.java
@@ -4,11 +4,7 @@ import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
-import android.support.v7.app.AppCompatActivity;
-import android.view.LayoutInflater;
 import android.view.View;
-import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.ImageView;
@@ -21,23 +17,18 @@ import com.example.dmitry.a1c_client.android.fragments.MessageDialogFragment;
 import com.example.dmitry.a1c_client.android.fragments.MessageDialogFragment.MessageCallBack;
 import com.example.dmitry.a1c_client.android.fragments.QuestionDialogFragment;
 import com.example.dmitry.a1c_client.android.fragments.QuestionDialogFragment.AnswerCallBack;
-import com.example.dmitry.a1c_client.di.DaggerIncomeFragmentComponent;
-import com.example.dmitry.a1c_client.di.DaggerIncomeTaskComponent;
-import com.example.dmitry.a1c_client.di.IncomeListModule;
-import com.example.dmitry.a1c_client.di.IncomeTaskModule;
-import com.example.dmitry.a1c_client.di.MainComponent;
+import com.example.dmitry.a1c_client.di.income_task.DaggerIncomeTaskComponent;
+import com.example.dmitry.a1c_client.di.income_task.DaggerIncomeTaskViewComponent;
+import com.example.dmitry.a1c_client.di.income_task.IncomeTaskComponent;
+import com.example.dmitry.a1c_client.di.income_task.IncomeTaskViewModule;
 import com.example.dmitry.a1c_client.domain.entity.NomenclaturePosition;
 import com.example.dmitry.a1c_client.domain.entity.StoreMapObject;
-import com.example.dmitry.a1c_client.misc.CommonFilters;
 import com.example.dmitry.a1c_client.presentation.IncomeTaskPresenter;
 import com.example.dmitry.a1c_client.presentation.IncomeTaskView;
 
 import javax.inject.Inject;
 
 import butterknife.BindView;
-import butterknife.ButterKnife;
-
-import static com.example.dmitry.a1c_client.misc.CommonFilters.isValidNumber;
 
 /**
  * Created by Admin on 19.12.2016.
@@ -89,11 +80,10 @@ public class IncomeTaskActivity extends BaseActivity implements IncomeTaskView, 
 
     @Override
     protected void initDi(Bundle savedInstanceState) {
-        if(((MyApplication)getApplicationContext()).getIncomeTaskComponent() != null){
-            ((MyApplication)getApplicationContext()).getIncomeTaskComponent().inject(this);
-        }else {
-            ((MyApplication)getApplicationContext()).buildIncomeTaskComponent(this).inject(this);
-        }
+        IncomeTaskComponent taskComponent = ((MyApplication)getApplicationContext())
+                .getIncomeTaskComponent();
+        DaggerIncomeTaskViewComponent.builder().incomeTaskComponent(taskComponent)
+                .incomeTaskViewModule(new IncomeTaskViewModule(this)).build().inject(this);
     }
 
     @Override
