+++ app/src/main/java/com/example/dmitry/a1c_client/di/IncomeFragmentComponent.java
@@ -1,6 +1,7 @@
 package com.example.dmitry.a1c_client.di;
 
 import com.example.dmitry.a1c_client.android.IncomeListFragment;
+import com.example.dmitry.a1c_client.di.scopes.PerActivity;
 
 import dagger.Component;
