+++ app/src/main/java/com/example/dmitry/a1c_client/di/income_task/IncomeTaskModule.java
@@ -1,12 +1,11 @@
-package com.example.dmitry.a1c_client.di;
+package com.example.dmitry.a1c_client.di.income_task;
 
 import com.example.dmitry.a1c_client.data.IncomeTaskRepositoryImpl;
+import com.example.dmitry.a1c_client.di.scopes.PerActivity;
 import com.example.dmitry.a1c_client.domain.IncomeTaskRepository;
 import com.example.dmitry.a1c_client.domain.StateKeeper;
 import com.example.dmitry.a1c_client.domain.entity.IncomeTaskState;
 import com.example.dmitry.a1c_client.presentation.IncomeTaskView;
-
-import javax.inject.Singleton;
 
 import dagger.Module;
 import dagger.Provides;
@@ -16,27 +15,18 @@ import dagger.Provides;
  */
 @Module
 public class IncomeTaskModule {
-    private IncomeTaskView incomeTaskView;
-
-    public IncomeTaskModule(IncomeTaskView incomeTaskView) {
-        this.incomeTaskView = incomeTaskView;
-    }
 
     @Provides
-    @Singleton
+    @PerActivity
     public StateKeeper<IncomeTaskState> provideIncomeTaskState(){
         return new StateKeeper<>(IncomeTaskState.EMPTY);
     }
 
     @Provides
-    @Singleton
+    @PerActivity
     public IncomeTaskRepository provideIncomeTaskRepository(){
         return new IncomeTaskRepositoryImpl();
     }
 
-    @Provides
-    @Singleton
-    public IncomeTaskView provideIncomeTaskView(){
-        return incomeTaskView;
-    }
+
 }
