+++ app/src/main/java/com/example/dmitry/a1c_client/di/scopes/PerActivity.java
@@ -1,0 +1,11 @@
+package com.example.dmitry.a1c_client.di.scopes;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+
+import javax.inject.Scope;
+
+@Scope
+@Retention(RetentionPolicy.RUNTIME)
+public @interface PerActivity {
+}
