+++ app/src/main/java/com/example/dmitry/a1c_client/di/scopes/PerInstance.java
@@ -1,11 +1,14 @@
-package com.example.dmitry.a1c_client.di;
+package com.example.dmitry.a1c_client.di.scopes;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 
 import javax.inject.Scope;
 
+/**
+ * Created by roma on 19.12.2016.
+ */
 @Scope
 @Retention(RetentionPolicy.RUNTIME)
-public @interface PerActivity {
+public @interface PerInstance {
 }
